Update: add comments and move functions to files to comply with the norm.
